Fidget around with roles and access control policies
Turn the "role" concept from a bitmask to a first-class object
in the database. A role becomes one of the different ways in which
a user can sign up as (for example as "patient" or as "researcher").

The role is then turned into a configurable set of capabilities,
which are then proxied to almond-dialog-agent using pluggable
policy modules.